refactor: 예약 응답 Dto 수정
